+++ src/Items/DanceLightItem.java
@@ -24,6 +24,6 @@ public class DanceLightItem extends SpecialItem{
 
     @Override
     public void draw(Graphics g) {
-        g.drawImage(Game.danceLight ,x ,y ,BricksBreaker.brickWidth ,BricksBreaker.brickHeight ,null ,null);
+        g.drawImage(Game.danceLight ,(int) x ,(int) y ,BricksBreaker.brickWidth ,BricksBreaker.brickHeight ,null ,null);
     }
 }
