+++ src/Items/OIG.java
@@ -5,24 +5,24 @@ import java.util.ArrayList;
 
 public abstract class OIG {
 
-    int x;
-    int y;
+    double x;
+    double y;
     double xVelocity;
     double yVelocity;
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
