+++ src/Items/PowerItem.java
@@ -36,6 +36,6 @@ public class PowerItem extends OrdinaryItem{
     @Override
     public void draw(Graphics g) {
         g.setColor(BricksBreaker.powerItemColor);
-        g.fillOval(x - BricksBreaker.itemRadios ,y - BricksBreaker.itemRadios ,BricksBreaker.itemRadios * 2 ,BricksBreaker.itemRadios * 2);
+        g.fillOval((int) x - BricksBreaker.itemRadios ,(int) y - BricksBreaker.itemRadios ,BricksBreaker.itemRadios * 2 ,BricksBreaker.itemRadios * 2);
     }
 }
