+++ src/Items/SpeedItem.java
@@ -43,6 +43,6 @@ public class SpeedItem extends OrdinaryItem{
     @Override
     public void draw(Graphics g) {
         g.setColor(BricksBreaker.speedItemColor);
-        g.fillOval(x - BricksBreaker.itemRadios ,y - BricksBreaker.itemRadios ,BricksBreaker.itemRadios * 2 ,BricksBreaker.itemRadios * 2);
+        g.fillOval((int) x - BricksBreaker.itemRadios ,(int) y - BricksBreaker.itemRadios ,BricksBreaker.itemRadios * 2 ,BricksBreaker.itemRadios * 2);
     }
 }
